type fix dbuf_cache_max_shift

+++ include/sys/kstat_osx.h
@@ -250,7 +250,7 @@ extern uint64_t send_holes_without_birth_time;
 extern uint64_t zfs_send_holes_without_birth_time;
 
 extern uint64_t dbuf_cache_max_bytes;
-extern uint64_t dbuf_cache_max_shift;
+extern int dbuf_cache_max_shift;
 
 extern uint64_t zfs_vdev_queue_depth_pct;
 extern boolean_t zio_dva_throttle_enabled;
